Fix request path access
When cherry-picking in the change to use swagger-ui-dist the if statment for
filtering out the dist's index.html file accesses the request path to see
what file is being requested. This came from swagger-ui-express and thus used
req.path; however, in Restify path is a function _not_ a property.

This change set updates the check to use req.path() instead of req.path to
make it work in Restify.